[deckhouse] extract package metadata mapping into a shared package
- `metadata.FromPackageDefinition` projects a parsed package.yaml onto the version status; the module-package-version controller now delegates to it.
- One mapping serves every writer of that status: the controller filling drafts from the registry and the startup sync filling embedded versions from disk.
- The legacy module.yaml pair stays in the controller: moving it would drag a moduleloader types dependency into internal.
